@@ -12,45 +12,45 @@
     "dev:netlify": "cross-env NODE_ENV=development netlify dev"
   },
   "dependencies": {
-    "@balavishnuvj/remix-seo": "^1.0.1",
+    "@balavishnuvj/remix-seo": "^1.0.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-brands-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.16",
+    "@fortawesome/react-fontawesome": "^0.1.19",
     "@kentcdodds/md-temp": "^1.1.0",
     "@netlify/functions": "^0.7.2",
-    "@prisma/client": "^3.7.0",
+    "@prisma/client": "^3.15.2",
     "@reach/menu-button": "^0.16.2",
-    "@react-spring/web": "^9.3.1",
-    "@remix-run/netlify": "^1.0.6",
-    "@remix-run/react": "^1.0.6",
-    "@tailwindcss/typography": "^0.4.1",
+    "@react-spring/web": "^9.4.5",
+    "@remix-run/netlify": "^1.6.4",
+    "@remix-run/react": "^1.6.4",
+    "@tailwindcss/typography": "^0.5.3",
     "@types/mailgun-js": "^0.22.12",
-    "clsx": "^1.1.1",
-    "esbuild": "^0.12.9",
-    "hamburger-react": "^2.4.1",
+    "clsx": "^1.2.1",
+    "esbuild": "^0.12.29",
+    "hamburger-react": "^2.5.0",
     "mailgun-js": "^0.22.0",
     "mdx-bundler": "^9.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^18.2.0",
-    "remix": "^1.0.6",
+    "remix": "^1.6.4",
     "tiny-invariant": "^1.2.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.0.6",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
+    "@remix-run/dev": "^1.6.4",
+    "@types/react": "^17.0.47",
+    "@types/react-dom": "^17.0.17",
     "@types/tailwindcss": "^2.2.4",
-    "autoprefixer": "^10.4.0",
-    "concurrently": "^6.4.0",
+    "autoprefixer": "^10.4.7",
+    "concurrently": "^6.5.1",
     "cross-env": "^7.0.3",
-    "postcss": "^8.4.4",
-    "postcss-cli": "^9.0.2",
-    "prisma": "^3.7.0",
-    "tailwindcss": "^2.2.19",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.1.2"
+    "postcss": "^8.4.14",
+    "postcss-cli": "^9.1.0",
+    "prisma": "^3.15.2",
+    "tailwindcss": "^3.1.5",
+    "ts-node": "^10.8.2",
+    "typescript": "^4.7.4"
   },
   "engines": {
     "node": ">=14"
