@@ -1,6 +1,6 @@
 import { renderToString } from "react-dom/server";
-import { RemixServer } from "remix";
-import type { EntryContext } from "remix";
+import { RemixServer } from "@remix-run/react";
+import type { EntryContext } from "@remix-run/node";
 import { otherRootRouteHandlers } from "~/utils/otherRootRoutes.server";
 
 export default async function handleRequest(
