@@ -1,5 +1,5 @@
 import { Post } from "@prisma/client";
-import { Link, useLoaderData } from "remix";
+import { useLoaderData } from "@remix-run/react";
 import PostCard from "~/components/PostCard";
 import { H1 } from "~/components/Typography";
 import { getPosts } from "~/utils/posts/posts.server";
@@ -14,7 +14,7 @@ export default function Posts() {
       <div className="remix__page">
         <div>
           <H1 className="py-5">Posts</H1>
-          <div className="md:grid md:grid-cols-2 md:gap-4">
+          <div className="md:grid md:grid-cols-2 md:gap-4 pb-5">
             {posts.map((post: Post) => (
               <PostCard key={post.slug} post={post} />
             ))}
