@@ -14,14 +14,17 @@ export async function getPosts(): Promise<Post[]> {
 
 export async function getPost(slug: string) {
   const post = await fetchPost(slug);
-  return {
-    slug: post.slug,
-    title: post.title,
-    code: post.body,
-    date: post.createdAt,
-    views: post.views,
-    img: post.img,
-  };
+  if (post) {
+    return {
+      slug: post?.slug,
+      title: post?.title,
+      code: post?.body,
+      date: post?.createdAt,
+      views: post?.views,
+      img: post?.img,
+    };
+  }
+  return undefined;
 }
 
 export async function incrementPostViews(slug: string) {
