@@ -77,7 +77,7 @@ export const handler: Handler = async (event, context) => {
       },
     })
     .then((resp) => {
-      console.log("update success", resp);
+      console.log("update success");
       return {
         statusCode: 200,
         body: JSON.stringify(resp),
